AVS-421_67 - Revise for Magento 2.2 compatibility
- Refactor code to prevent DI compilation errors
